Revert "refactor: split proxy & server starting logic (#36)"
This reverts commit 718afb383037481941d9a2502e17ade2fd781f7c.
