+++ cmd/cmd.go
@@ -8,7 +8,7 @@ import (
 
 func New(logger log.Logger, appConfig *config.Shield) *cli.Command {
 	var cmd = &cli.Command{
-		Use:          "shield",
+		Use:          "proxy",
 		SilenceUsage: true,
 	}
